+++ MTCG_Project/MTCG/Cards/Monsters/FireElf.cs
@@ -6,7 +6,7 @@ namespace MTCG_Project.MTCG.Cards.Monsters
 {
     public class FireElf : AbstractMonstercard
     {
-        public FireElf(int damage) : base(Cardname.FireElf, damage) { }
+        public FireElf(float damage) : base(Cardname.FireElf, damage) { }
 
         public override bool SpecialBehavior(ICard otherCard)
         {
